+++ not1mm/plugins/darc_vhf.py
@@ -147,8 +147,9 @@ def prefill(self):
     """Fill SentNR"""
     sent_sxchange_setting = self.contest_settings.get("SentExchange", "")
     if sent_sxchange_setting.strip() == "#":
-        result = self.database.get_serial()
-        serial_nr = str(result.get("serial_nr", "1")).zfill(3)
+        # result = self.database.get_serial()
+        # serial_nr = str(result.get("serial_nr", "1")).zfill(3)
+        serial_nr = str(self.current_sn).zfill(3)
         if serial_nr == "None":
             serial_nr = "001"
         if len(self.other_1.text()) == 0: