+++ not1mm/plugins/ea_rtty.py
@@ -161,8 +161,9 @@ def prefill(self) -> None:
     """
     Fill the SentNR field with either the next serial number or the province.
     """
-    result = self.database.get_serial()
-    serial_nr = str(result.get("serial_nr", "1")).zfill(3)
+    # result = self.database.get_serial()
+    # serial_nr = str(result.get("serial_nr", "1")).zfill(3)
+    serial_nr = str(self.current_sn).zfill(3)
     if serial_nr == "None":
         serial_nr = "001"