+++ not1mm/plugins/jidx_ph.py
@@ -108,8 +108,9 @@ def prefill(self):
     """Fill SentNR"""
     SentExchange = self.contest_settings.get("SentExchange", "")
     if "#" in SentExchange:
-        result = self.database.get_serial()
-        serial_nr = str(result.get("serial_nr", "1")).zfill(3)
+        # result = self.database.get_serial()
+        # serial_nr = str(result.get("serial_nr", "1")).zfill(3)
+        serial_nr = str(self.current_sn).zfill(3)
         if serial_nr == "None":
             serial_nr = "001"
         SentExchange.replace("#", str(serial_nr))