+++ not1mm/plugins/ukeidx.py
@@ -113,8 +113,9 @@ def predupe(self):
 def prefill(self):
     """Fill sentnr"""
     exch = str(self.contest_settings.get("SentExchange", 0))
-    result = self.database.get_serial()
-    serial_nr = str(result.get("serial_nr", "1")).zfill(3)
+    # result = self.database.get_serial()
+    # serial_nr = str(result.get("serial_nr", "1")).zfill(3)
+    serial_nr = str(self.current_sn).zfill(3)
     if serial_nr == "None":
         serial_nr = "001"
     if len(self.other_1.text()) == 0: